render: widen trailMaxIndex/trailNextIndex to s32 to drop u8 masks
The trail history indices were u8, forcing clrlwi/cntlzw byte-mask idioms
on every increment and equality test. The target keeps them as ints and
uses plain cmpw, masking only at the history[] index. render 59.07->59.99%,
unit 82.78->83.00%.

+++ src/pppYmMegaBirthShpTail2.cpp
@@ -116,8 +116,8 @@ void pppRenderYmMegaBirthShpTail2(pppYmMegaBirthShpTail2* object, pppYmMegaBirth
             tagOAN3_SHAPE* shape =
                 reinterpret_cast<tagOAN3_SHAPE*>(reinterpret_cast<u8*>(shapeAnim) + shapeFrame->m_shapeOffset);
             const u8 trailReadIndex = *(u8*)(particle + 0x38);
-            const u8 trailMaxIndex = (u8)(*(u8*)(particle + 0x37) - 1);
-            u8 trailNextIndex = (u8)(trailReadIndex + 1);
+            const s32 trailMaxIndex = (u8)(*(u8*)(particle + 0x37) - 1);
+            s32 trailNextIndex = (u8)(trailReadIndex + 1);
             const float alphaScale = (float)*(s16*)((u8*)colorWork + 6) / kPppYmMegaBirthShpTail2AlphaDivisor;
             const float stepDivisor = (float)((s32)frameCountRaw - 1);
             float drawScale = *(float*)(step + 0x70);
